Add provider type and id to canonical_facts
