+++ app/models.py
@@ -26,7 +26,6 @@ from sqlalchemy.dialects.postgresql import UUID
 from yaml import safe_load
 
 from app.exceptions import InventoryException
-from app.exceptions import ValidationException
 from app.logging import get_logger
 from app.validators import check_empty_keys
 from app.validators import verify_satellite_id
@@ -45,14 +44,6 @@ SPECIFICATION_DIR = "./swagger/"
 SYSTEM_PROFILE_SPECIFICATION_FILE = "system_profile.spec.yaml"
 
 
-class ProviderType(str, Enum):
-    ALIBABA = "alibaba"
-    AWS = "aws"
-    AZURE = "azure"
-    GCP = "gcp"
-    IBM = "ibm"
-
-
 def _set_display_name_on_save(context):
     """
     This method sets the display_name if it has not been set previously.
@@ -62,18 +53,6 @@ def _set_display_name_on_save(context):
     params = context.get_current_parameters()
     if not params["display_name"]:
         return params["canonical_facts"].get("fqdn") or params["id"]
-
-
-#  verifies provider_type and if the required provider_id is provided.
-def _check_provider(provider_type, canonical_facts):
-    if provider_type.lower() not in ProviderType.__members__.values():
-        raise ValidationException(
-            f'Unknown Provider Type: {provider_type}.  Valid provider types are:\
-             "alibaba", "aws", "azure", "gcp", or "ibm"'
-        )
-    if not canonical_facts.get("provider_id"):
-        raise ValidationException("Missing Provider ID")
-    return True
 
 
 def _time_now():
@@ -165,7 +144,6 @@ class LimitedHost(db.Model):
         facts=None,
         tags=None,
         system_profile_facts=None,
-        provider_type=None,
     ):
 
         self.canonical_facts = canonical_facts
@@ -180,17 +158,11 @@ class LimitedHost(db.Model):
         self.facts = facts or {}
         self.tags = tags or {}
         self.system_profile_facts = system_profile_facts or {}
-        self._update_provider_type(provider_type)
 
     def _update_ansible_host(self, ansible_host):
         if ansible_host is not None:
             # Allow a user to clear out the ansible host with an empty string
             self.ansible_host = ansible_host
-
-    def _update_provider_type(self, provider_type):
-        if provider_type is not None:
-            # Allow a user to clear out the provider_type with an empty string
-            self.provider_type = provider_type
 
     id = db.Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     account = db.Column(db.String(10))
@@ -202,7 +174,6 @@ class LimitedHost(db.Model):
     tags = db.Column(JSONB)
     canonical_facts = db.Column(JSONB)
     system_profile_facts = db.Column(JSONB)
-    provider_type = db.Column(db.String(50))
 
 
 class Host(LimitedHost):
@@ -221,7 +192,6 @@ class Host(LimitedHost):
         system_profile_facts=None,
         stale_timestamp=None,
         reporter=None,
-        provider_type=None,
     ):
         if not canonical_facts:
             raise InventoryException(
@@ -233,12 +203,8 @@ class Host(LimitedHost):
                 title="Invalid request", detail="Both stale_timestamp and reporter fields must be present."
             )
 
-        if provider_type:
-            _check_provider(provider_type, canonical_facts)
-
-        super().__init__(
-            canonical_facts, display_name, ansible_host, account, facts, tags, system_profile_facts, provider_type
-        )
+        # )
+        super().__init__(canonical_facts, display_name, ansible_host, account, facts, tags, system_profile_facts)
         self.stale_timestamp = stale_timestamp
         self.reporter = reporter
         self._update_per_reporter_staleness(stale_timestamp, reporter)
@@ -264,8 +230,6 @@ class Host(LimitedHost):
         self._update_stale_timestamp(input_host.stale_timestamp, input_host.reporter)
         self._update_per_reporter_staleness(input_host.stale_timestamp, input_host.reporter)
 
-        self._update_provider_type(input_host.provider_type)
-
     def patch(self, patch_data):
         logger.debug("patching host (id=%s) with data: %s", self.id, patch_data)
 
@@ -275,8 +239,6 @@ class Host(LimitedHost):
         self._update_ansible_host(patch_data.get("ansible_host"))
 
         self.update_display_name(patch_data.get("display_name"))
-
-        self._update_provider_type(patch_data.get("provider_type"))
 
     def update_display_name(self, input_display_name):
         if input_display_name:
@@ -461,6 +423,7 @@ class CanonicalFactsSchema(MarshmallowSchema):
     )
     external_id = fields.Str(validate=marshmallow_validate.Length(min=1, max=500))
     provider_id = fields.Str(validate=marshmallow_validate.Length(min=1, max=50))
+    provider_type = fields.Str(validate=marshmallow_validate.Length(min=1, max=50))
 
 
 class LimitedHostSchema(CanonicalFactsSchema):
@@ -470,7 +433,6 @@ class LimitedHostSchema(CanonicalFactsSchema):
     facts = fields.List(fields.Nested(FactsSchema))
     system_profile = fields.Dict()
     tags = fields.Raw(allow_none=True)
-    provider_type = fields.Str(validate=marshmallow_validate.Length(min=0, max=50))
 
     def __init__(self, system_profile_schema=None, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -540,7 +502,6 @@ class LimitedHostSchema(CanonicalFactsSchema):
             facts=facts,
             tags=tags,
             system_profile_facts=data.get("system_profile", {}),
-            provider_type=data.get("provider_type"),
         )
 
     @pre_load
@@ -579,7 +540,6 @@ class HostSchema(LimitedHostSchema):
             data.get("system_profile", {}),
             data["stale_timestamp"],
             data["reporter"],
-            data.get("provider_type"),
         )
 
     @validates("stale_timestamp")
@@ -591,7 +551,6 @@ class HostSchema(LimitedHostSchema):
 class PatchHostSchema(MarshmallowSchema):
     ansible_host = fields.Str(validate=marshmallow_validate.Length(min=0, max=255))
     display_name = fields.Str(validate=marshmallow_validate.Length(min=1, max=200))
-    provider_type = fields.Str(validate=marshmallow_validate.Length(min=0, max=50))
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
