+++ tests/test_unit.py
@@ -912,7 +912,6 @@ class SerializationDeserializeHostMockedTestCase(TestCase):
             "fqdn": "some fqdn",
             "mac_addresses": ["c2:00:d0:c8:61:01"],
             "external_id": "i-05d2313e6b9a42b16",
-            "provider_type": "aws",
             "facts": {
                 "some namespace": {"some key": "some value"},
                 "another namespace": {"another key": "another value"},
@@ -949,7 +948,6 @@ class SerializationDeserializeHostMockedTestCase(TestCase):
             host_input["system_profile"],
             host_input["stale_timestamp"],
             host_input["reporter"],
-            host_input["provider_type"],
         )
 
     def test_without_facts(self, deserialize_canonical_facts, deserialize_facts, deserialize_tags, host):
@@ -969,7 +967,6 @@ class SerializationDeserializeHostMockedTestCase(TestCase):
             },
             "stale_timestamp": datetime.now(timezone.utc).isoformat(),
             "reporter": "some reporter",
-            "provider_type": "aws",
         }
         host_schema = Mock(
             **{"return_value.load.return_value.data": host_input, "build_model": HostSchema.build_model}
@@ -991,7 +988,6 @@ class SerializationDeserializeHostMockedTestCase(TestCase):
             host_input["system_profile"],
             host_input["stale_timestamp"],
             host_input["reporter"],
-            host_input["provider_type"],
         )
 
     def test_without_tags(self, deserialize_canonical_facts, deserialize_facts, deserialize_tags, host):
@@ -1011,7 +1007,6 @@ class SerializationDeserializeHostMockedTestCase(TestCase):
             },
             "stale_timestamp": datetime.now(timezone.utc).isoformat(),
             "reporter": "some reporter",
-            "provider_type": "aws",
         }
         host_schema = Mock(
             **{"return_value.load.return_value.data": host_input, "build_model": HostSchema.build_model}
@@ -1033,7 +1028,6 @@ class SerializationDeserializeHostMockedTestCase(TestCase):
             host_input["system_profile"],
             host_input["stale_timestamp"],
             host_input["reporter"],
-            host_input["provider_type"],
         )
 
     def test_without_display_name(self, deserialize_canonical_facts, deserialize_facts, deserialize_tags, host):
@@ -1056,7 +1050,6 @@ class SerializationDeserializeHostMockedTestCase(TestCase):
             },
             "stale_timestamp": datetime.now(timezone.utc).isoformat(),
             "reporter": "some reporter",
-            "provider_type": "aws",
         }
         host_schema = Mock(
             **{"return_value.load.return_value.data": host_input, "build_model": HostSchema.build_model}
@@ -1078,7 +1071,6 @@ class SerializationDeserializeHostMockedTestCase(TestCase):
             host_input["system_profile"],
             host_input["stale_timestamp"],
             host_input["reporter"],
-            host_input["provider_type"],
         )
 
     def test_without_system_profile(self, deserialize_canonical_facts, deserialize_facts, deserialize_tags, host):
@@ -1096,7 +1088,6 @@ class SerializationDeserializeHostMockedTestCase(TestCase):
             },
             "stale_timestamp": datetime.now(timezone.utc).isoformat(),
             "reporter": "some reporter",
-            "provider_type": "aws",
         }
         host_schema = Mock(
             **{"return_value.load.return_value.data": host_input, "build_model": HostSchema.build_model}
@@ -1118,7 +1109,6 @@ class SerializationDeserializeHostMockedTestCase(TestCase):
             {},
             host_input["stale_timestamp"],
             host_input["reporter"],
-            host_input["provider_type"],
         )
 
     def test_without_culling_fields(self, deserialize_canonical_facts, deserialize_facts, deserialize_tags, host):
@@ -1212,13 +1202,13 @@ class SerializationSerializeHostCompoundTestCase(SerializationSerializeHostBaseT
             "mac_addresses": ["c2:00:d0:c8:61:01"],
             "external_id": "i-05d2313e6b9a42b16",
             "provider_id": "i-05d2313e6b9a42b16",
+            "provider_type": "aws",
         }
         unchanged_data = {
             "display_name": "some display name",
             "ansible_host": "some ansible host",
             "account": "some acct",
             "reporter": "insights",
-            "provider_type": "aws",
         }
         host_init_data = {
             "canonical_facts": canonical_facts,
@@ -1394,7 +1384,6 @@ class SerializationSerializeHostMockedTestCase(SerializationSerializeHostBaseTes
             "stale_timestamp": self._timestamp_to_str(staleness_offset.stale_timestamp.return_value),
             "stale_warning_timestamp": self._timestamp_to_str(staleness_offset.stale_warning_timestamp.return_value),
             "culled_timestamp": self._timestamp_to_str(staleness_offset.culled_timestamp.return_value),
-            "provider_type": None,
         }
         self.assertEqual(expected, actual)
 
@@ -1511,6 +1500,7 @@ class SerializationSerializeCanonicalFactsTestCase(TestCase):
             "mac_addresses": ("c2:00:d0:c8:61:01",),
             "external_id": "i-05d2313e6b9a42b16",
             "provider_id": "i-05d2313e6b9a42b16",
+            "provider_type": "aws",
         }
         self.assertEqual(canonical_facts, serialize_canonical_facts(canonical_facts))
 
@@ -1526,6 +1516,7 @@ class SerializationSerializeCanonicalFactsTestCase(TestCase):
             "mac_addresses",
             "external_id",
             "provider_id",
+            "provider_type",
         )
         self.assertEqual({field: None for field in canonical_fact_fields}, serialize_canonical_facts({}))
